Modify code for linters errors
